[DDP] support arbitrary output type by ArgsTree
