+++ python/oneflow/utils/global_view/to_global.py
@@ -19,7 +19,7 @@ import types
 
 import oneflow as flow
 from oneflow.framework.tensor import Tensor
-from oneflow.nn.graph.util import ArgsTree
+from oneflow.framework.args_tree import ArgsTree
 from oneflow.utils.global_view.global_utils import (
     to_global_tensor,
     check_input_global,
